Improved readability of path follower with comments, field renaming and and code restructuring.
Consolidated crosswalk management algorithm.

The robot now compares its results to plates.csv in the scripts directory to determine success or failure.

main algorithm renamed parking_bot.py

main algorithm now run out of my_launch.launch
