+++ launch/my_launch.launch
@@ -1,6 +1,6 @@
 <launch>
     <param name="/use_sim_time" value="true"/>
-    <arg name="spawn_pedestrians" default="false"/>
+    <arg name="spawn_pedestrians" default="True"/>
 
     <include file="$(find gazebo_ros)/launch/empty_world.launch">
         <env name="GAZEBO_RESOURCE_PATH" value="$(find competition_2019t2)"/>
@@ -16,4 +16,12 @@
     <include file="$(find competition_2019t2)/launch/npcs.launch">
         <arg name="spawn_pedestrians" value="$(arg spawn_pedestrians)"/>
     </include>
+
+    <!-- Package launch file -->
+    <node
+            pkg="competition_2019t2"
+            type="parking_bot.py"
+            name="parking_bot"
+            output="screen">
+    </node>
 </launch>
