+++ nodes/assets/homography-sim_v4.npy
