+++ nodes/assets/shape-sim_v0.npy
