+++ nodes/assets/shape-sim_v1.npy
