+++ nodes/assets/shape-sim_v2.npy
