+++ nodes/assets/shape-sim_v4.npy
