+++ nodes/global_variables.py
