+++ nodes/plate_isolator_colour.py
