+++ scripts/plates.csv
@@ -1,16 +1,16 @@
-UT95
-YS21
-AY44
-AL32
-NJ22
-BL83
-BT90
-OP25
-LB99
-JD90
-QF57
-RO00
-UX24
-IS40
-FL98
-UL50
+EK24
+HE58
+GX23
+FQ53
+WO73
+VD06
+NH90
+FV44
+YO78
+CG67
+SS39
+AY31
+TW08
+PX05
+RU98
+YF57
